+++ R/allGenerics.R
@@ -2,7 +2,7 @@
 #' @rdname combineTransitions
 #'
 #' @param object a SRM object
-#' @return a list of SRM total ion chromatograms
+#' @return a \code{transition} (list) class of SRM total ion chromatograms
 #'
 #' @author Tom Wilson \email{tpw2@@aber.ac.uk}
 #' @export
@@ -39,4 +39,20 @@ setGeneric(name = "meta",
 setGeneric(name = "transitions",
            def = function(object)
            {standardGeneric("transitions")}
+)
+
+
+
+#' Plot SRM
+#' @rdname plotSRM
+#' @param object a SRM objecy
+#' @param idn the index number of a transition to plot
+#' @return NULL
+#'
+#' @author Tom Wilson \email{tpw2@@aber.ac.uk}
+#' @export
+
+setGeneric(name = "plotSRM",
+           def = function(object,idn)
+           {standardGeneric("plotSRM")}
 )